requirements.txt
flask
sounddevice
numpy
wave
google-generativeai